+++ crates/sml-statics/src/dec.rs
@@ -144,7 +144,7 @@ pub(crate) fn get(
             if let Some(ref param) = param {
               ty = Ty::fun(param.clone(), ty);
             }
-            let exn = st.syms.insert_exn(name.clone(), param);
+            let exn = st.syms.insert_exn(st.mk_path(name.clone()), param);
             let vi = ValInfo {
               ty_scheme: TyScheme::zero(ty),
               id_status: IdStatus::Exn(exn),
@@ -290,7 +290,7 @@ pub(crate) fn get_dat_binds(
   // do a first pass through the datatypes to allow for recursive reference, and to put them in
   // scope for the types.
   for dat_bind in dat_binds.iter() {
-    let started = st.syms.start(dat_bind.name.clone());
+    let started = st.syms.start(st.mk_path(dat_bind.name.clone()));
     // just create the fixed ty vars, do not bring them into the scope of the cx yet.
     let mut fixed = FixedTyVars::default();
     for ty_var in &dat_bind.ty_vars {